requirements.txt
# llm-commander/requirements.txt
python-dotenv~=1.1.0
google-generativeai
Flask~=3.1.0
Flask-Login~=0.6.3
Flask-WTF
requests
werkzeug~=3.1.3
pexpect~=4.9.0
gunicorn
protobuf~=5.29.4
WTForms~=3.2.1